+++ includes/meta_utils.hpp
@@ -38,38 +38,38 @@ using namespace std;
 
 // metafunctions to get the tuple type from a callable type (e.g., function, lambda, functor)
 template<typename F_t, typename Ret, typename Arg1, typename Arg2>
-Arg1 get_tuple_t(Ret (F_t::*)(size_t, size_t, Iterable<Arg1>&, Arg2&) const);
+Arg1 get_tuple_t(Ret (F_t::*)(size_t, uint64_t, Iterable<Arg1>&, Arg2&) const);
 
 template<typename Ret, typename Arg1, typename Arg2>
-Arg1 get_tuple_t(Ret (*)(size_t, size_t, Iterable<Arg1>&, Arg2&));
+Arg1 get_tuple_t(Ret (*)(size_t, uint64_t, Iterable<Arg1>&, Arg2&));
 
 template<typename F_t, typename Ret, typename Arg1, typename Arg2>
-Arg1 get_tuple_t(Ret (F_t::*)(size_t, size_t, const Arg1&, Arg2&) const);
+Arg1 get_tuple_t(Ret (F_t::*)(size_t, uint64_t, const Arg1&, Arg2&) const);
 
 template<typename Ret, typename Arg1, typename Arg2>
-Arg1 get_tuple_t(Ret (*)(size_t, size_t, const Arg1&, Arg2&));
+Arg1 get_tuple_t(Ret (*)(size_t, uint64_t, const Arg1&, Arg2&));
 
 template<typename F_t, typename Ret, typename Arg1, typename Arg2>
-Arg1 get_tuple_t(Ret (F_t::*)(size_t, size_t, const Arg1*, Arg2*, size_t, char*) const);
+Arg1 get_tuple_t(Ret (F_t::*)(size_t, uint64_t, const Arg1*, Arg2*, size_t, char*) const);
 
 template<typename Ret, typename Arg1, typename Arg2>
-Arg1 get_tuple_t(Ret (*)(size_t, size_t, const Arg1*, Arg2*, size_t, char*));
+Arg1 get_tuple_t(Ret (*)(size_t, uint64_t, const Arg1*, Arg2*, size_t, char*));
 
 template<typename F_t>
 decltype(get_tuple_t(&F_t::operator())) get_tuple_t(F_t);
 
 // metafunctions to get the result type from a callable type (e.g., function, lambda, functor)
 template<typename F_t, typename Ret, typename Arg1, typename Arg2>
-Arg2 get_result_t(Ret (F_t::*)(size_t, size_t, Arg1&, Arg2&) const);
+Arg2 get_result_t(Ret (F_t::*)(size_t, uint64_t, Arg1&, Arg2&) const);
 
 template<typename Ret, typename Arg1, typename Arg2>
-Arg2 get_result_t(Ret (*)(size_t, size_t, Arg1&, Arg2&));
+Arg2 get_result_t(Ret (*)(size_t, uint64_t, Arg1&, Arg2&));
 
 template<typename F_t, typename Ret, typename Arg1, typename Arg2>
-Arg2 get_result_t(Ret (F_t::*)(size_t, size_t, const Arg1*, Arg2*, size_t, char*) const);
+Arg2 get_result_t(Ret (F_t::*)(size_t, uint64_t, const Arg1*, Arg2*, size_t, char*) const);
 
 template<typename Ret, typename Arg1, typename Arg2>
-Arg2 get_result_t(Ret (*)(size_t, size_t, const Arg1*, Arg2*, size_t, char*));
+Arg2 get_result_t(Ret (*)(size_t, uint64_t, const Arg1*, Arg2*, size_t, char*));
 
 template<typename F_t>
 decltype(get_result_t(&F_t::operator())) get_result_t(F_t);
@@ -142,28 +142,28 @@ auto get_KF_GPU_nested_type(F_t _f)
 
 // metafunctions to return the callable type to be executed on the GPU (only lambda or functor!)
 template<typename F_t, typename G_t, typename Ret, typename Arg1, typename Arg2>
-F_t get_GPU_F(Ret (F_t::*)(size_t, size_t, const Arg1*, Arg2*, size_t, char*) const, Ret (G_t::*)(size_t, size_t, Iterable<Arg2>&, Arg2&) const);
+F_t get_GPU_F(Ret (F_t::*)(size_t, uint64_t, const Arg1*, Arg2*, size_t, char*) const, Ret (G_t::*)(size_t, uint64_t, Iterable<Arg2>&, Arg2&) const);
 
 template<typename F_t, typename Ret, typename Arg1, typename Arg2>
-F_t get_GPU_F(Ret (F_t::*)(size_t, size_t, const Arg1*, Arg2*, size_t, char*) const, Ret (*)(size_t, size_t, Iterable<Arg2>&, Arg2&));
+F_t get_GPU_F(Ret (F_t::*)(size_t, uint64_t, const Arg1*, Arg2*, size_t, char*) const, Ret (*)(size_t, uint64_t, Iterable<Arg2>&, Arg2&));
 
 template<typename F_t, typename G_t, typename Ret, typename Arg1, typename Arg2>
-F_t get_GPU_F(Ret (F_t::*)(size_t, size_t, const Arg1*, Arg2*, size_t, char*) const, Ret (G_t::*)(size_t, size_t, const Arg2&, Arg2&) const);
+F_t get_GPU_F(Ret (F_t::*)(size_t, uint64_t, const Arg1*, Arg2*, size_t, char*) const, Ret (G_t::*)(size_t, uint64_t, const Arg2&, Arg2&) const);
 
 template<typename F_t, typename Ret, typename Arg1, typename Arg2>
-F_t get_GPU_F(Ret (F_t::*)(size_t, size_t, const Arg1*, Arg2*, size_t, char*) const, Ret (*)(size_t, size_t, const Arg2&, Arg2&));
+F_t get_GPU_F(Ret (F_t::*)(size_t, uint64_t, const Arg1*, Arg2*, size_t, char*) const, Ret (*)(size_t, uint64_t, const Arg2&, Arg2&));
 
 template<typename F_t, typename G_t, typename Ret, typename Arg1, typename Arg2>
-G_t get_GPU_F(Ret (F_t::*)(size_t, size_t, Iterable<Arg1>&, Arg2&) const, Ret(G_t::*)(size_t, size_t, const Arg2*, Arg2*, size_t, char*) const);
+G_t get_GPU_F(Ret (F_t::*)(size_t, uint64_t, Iterable<Arg1>&, Arg2&) const, Ret(G_t::*)(size_t, uint64_t, const Arg2*, Arg2*, size_t, char*) const);
 
 template<typename G_t, typename Ret, typename Arg1, typename Arg2>
-G_t get_GPU_F(Ret (*)(size_t, size_t, Iterable<Arg1>&, Arg2&), Ret(G_t::*)(size_t, size_t, const Arg2*, Arg2*, size_t, char*) const);
+G_t get_GPU_F(Ret (*)(size_t, uint64_t, Iterable<Arg1>&, Arg2&), Ret(G_t::*)(size_t, uint64_t, const Arg2*, Arg2*, size_t, char*) const);
 
 template<typename F_t, typename G_t, typename Ret, typename Arg1, typename Arg2>
-G_t get_GPU_F(Ret (F_t::*)(size_t, size_t, const Arg1&, Arg2&) const, Ret(G_t::*)(size_t, size_t, const Arg2*, Arg2*, size_t, char*) const);
+G_t get_GPU_F(Ret (F_t::*)(size_t, uint64_t, const Arg1&, Arg2&) const, Ret(G_t::*)(size_t, uint64_t, const Arg2*, Arg2*, size_t, char*) const);
 
 template<typename G_t, typename Ret, typename Arg1, typename Arg2>
-G_t get_GPU_F(Ret (*)(size_t, size_t, const Arg1&, Arg2&), Ret(G_t::*)(size_t, size_t, const Arg2*, Arg2*, size_t, char*) const);
+G_t get_GPU_F(Ret (*)(size_t, uint64_t, const Arg1&, Arg2&), Ret(G_t::*)(size_t, uint64_t, const Arg2*, Arg2*, size_t, char*) const);
 
 // wrapper struct of input tuples
 template<typename tuple_t>
@@ -200,7 +200,7 @@ void deleteTuple(typename enable_if<!is_same<T1,T2>::value, T2>::type *wt) // T1
 {
     T1 *t = wt->tuple;
     // check if the tuple and the wrapper must be destroyed/deallocated
-    int old_cnt = (wt->counter).fetch_sub(1);
+    size_t old_cnt = (wt->counter).fetch_sub(1);
     if (old_cnt == 1) {
         delete t;
         delete wt;
