+++ includes/pane_farm.hpp
@@ -60,13 +60,13 @@ class Pane_Farm: public ff_pipeline
 {
 private:
     // function type of the non-incremental pane processing
-    using f_plqfunction_t = function<int(size_t, size_t, Iterable<tuple_t> &, result_t &)>;
+    using f_plqfunction_t = function<int(size_t, uint64_t, Iterable<tuple_t> &, result_t &)>;
     // Function type of the incremental pane processing
-    using f_plqupdate_t = function<int(size_t, size_t, const tuple_t &, result_t &)>;
+    using f_plqupdate_t = function<int(size_t, uint64_t, const tuple_t &, result_t &)>;
     // function type of the non-incremental window processing
-    using f_wlqfunction_t = function<int(size_t, size_t, Iterable<result_t> &, result_t &)>;
+    using f_wlqfunction_t = function<int(size_t, uint64_t, Iterable<result_t> &, result_t &)>;
     // function type of the incremental window function
-    using f_wlqupdate_t = function<int(size_t, size_t, const result_t &, result_t &)>;
+    using f_wlqupdate_t = function<int(size_t, uint64_t, const result_t &, result_t &)>;
     // friendships with other classes in the library
     template<typename T1, typename T2, typename T3>
     friend class Win_Farm;
@@ -77,7 +77,7 @@ private:
     template<typename T>
     friend class KeyFarm_Builder;
     // compute the gcd between two numbers
-    function<size_t(size_t, size_t)> gcd = [](size_t u, size_t v) {
+    function<uint64_t(uint64_t, uint64_t)> gcd = [](uint64_t u, uint64_t v) {
         while (v != 0) {
             unsigned long r = u % v;
             u = v;
@@ -145,7 +145,7 @@ private:
             exit(EXIT_FAILURE);
         }
         // compute the pane length (no. of tuples or in time units)
-        size_t _pane_len = gcd(_win_len, _slide_len);
+        uint64_t _pane_len = gcd(_win_len, _slide_len);
         // general fastflow pointers to the PLQ and WLQ stages
         ff_node *plq_stage, *wlq_stage;
         // create the first stage PLQ (Pane Level Query)
@@ -223,7 +223,7 @@ private:
             exit(EXIT_FAILURE);
         }
         // compute the pane length (no. of tuples or in time units)
-        size_t _pane_len = gcd(_win_len, _slide_len);
+        uint64_t _pane_len = gcd(_win_len, _slide_len);
         // general fastflow pointers to the PLQ and WLQ stages
         ff_node *plq_stage, *wlq_stage;
         // create the first stage PLQ
@@ -301,7 +301,7 @@ private:
             exit(EXIT_FAILURE);
         }
         // compute the pane length (no. of tuples or in time units)
-        size_t _pane_len = gcd(_win_len, _slide_len);
+        uint64_t _pane_len = gcd(_win_len, _slide_len);
         // general fastflow pointers to the PLQ and WLQ stages
         ff_node *plq_stage, *wlq_stage;
         // create the first stage PLQ (Pane Level Query)
@@ -379,7 +379,7 @@ private:
             exit(EXIT_FAILURE);
         }
         // compute the pane length (no. of tuples or in time units)
-        size_t _pane_len = gcd(_win_len, _slide_len);
+        uint64_t _pane_len = gcd(_win_len, _slide_len);
         // general fastflow pointers to the PLQ and WLQ stages
         ff_node *plq_stage, *wlq_stage;
         // create the first stage PLQ (Pane Level Query)
